requirements.txt
Flask
pandas
numpy
matplotlib
seaborn
qrcode
Pillow
sqlalchemy
scipy
openpyxl
